Connector / Operator refactor in progress
