+++ arroyo-metrics/Cargo.toml
@@ -6,3 +6,4 @@ edition = "2021"
 [dependencies]
 arroyo-types = { path = "../arroyo-types" }
 prometheus = {version = "0.13", features = ["process"] }
+lazy_static = "1.4.0"
