+++ arroyo-operator/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "arroyo-operator"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+arroyo-formats = { path = "../arroyo-formats" }
+arroyo-metrics = { path = "../arroyo-metrics" }
+arroyo-rpc = { path = "../arroyo-rpc" }
+arroyo-state = { path = "../arroyo-state" }
+arroyo-types = { path = "../arroyo-types" }
+
+anyhow = "1.0.71"
+arrow = { workspace = true }
+ahash = "0.8"
+async-trait = "0.1.68"
+bincode = "2.0.0-rc.3"
+datafusion = "34.0"
+futures = "0.3"
+prost = "0.12"
+rand = "0.8"
+tokio = { version = "1", features = ["full", "tracing"] }
+tokio-stream = { version = "0.1", features = ["full"] }
+tracing = "0.1"
+async-stream = "0.3.5"
