+++ arroyo-rpc/Cargo.toml
@@ -26,6 +26,7 @@ async-trait = "0.1.74"
 apache-avro = "0.16.0"
 regex = "1.9.5"
 base64 = "0.21.5"
+ahash = "0.8.7"
 
 [build-dependencies]
 tonic-build = { workspace = true }
