+++ arroyo-state/src/schemas/mod.rs
@@ -9,8 +9,7 @@ use arrow_array::{
 };
 use arrow_ord::cmp::{gt_eq, lt_eq};
 use arrow_schema::{DataType, Field, Schema, SchemaRef};
-use arroyo_datastream::get_hasher;
-use arroyo_rpc::ArroyoSchema;
+use arroyo_rpc::{ArroyoSchema, get_hasher};
 use arroyo_types::from_nanos;
 use bincode::config;
 use datafusion_common::{hash_utils::create_hashes, ScalarValue};
