+++ arroyo-worker/src/arrow/session_aggregating_window.rs
@@ -30,8 +30,8 @@ use arroyo_types::{
 };
 use datafusion::{execution::context::SessionContext, physical_plan::ExecutionPlan};
 
-use crate::operator::{ArrowOperator, ArrowOperatorConstructor};
-use crate::{engine::ArrowContext, operator::OperatorNode};
+use arroyo_operator::operator::{ArrowOperator, OperatorConstructor};
+use arroyo_operator::{context::ArrowContext, operator::OperatorNode};
 use arroyo_df::physical::{ArroyoPhysicalExtensionCodec, DecodingContext};
 use datafusion_execution::{
     runtime_env::{RuntimeConfig, RuntimeEnv},
@@ -697,10 +697,9 @@ fn start_time_for_sorted_batch(batch: &RecordBatch, schema: &ArroyoSchema) -> Sy
     from_nanos(min_timestamp as u128)
 }
 
-impl ArrowOperatorConstructor<api::SessionWindowAggregateOperator>
-    for SessionAggregatingWindowFunc
-{
-    fn from_config(config: api::SessionWindowAggregateOperator) -> anyhow::Result<OperatorNode> {
+impl OperatorConstructor for SessionAggregatingWindowFunc {
+    type ConfigT = api::SessionWindowAggregateOperator;
+    fn with_config(&self, config: api::SessionWindowAggregateOperator) -> anyhow::Result<OperatorNode> {
         let window_field = Arc::new(Field::new(
             config.window_field_name,
             window_arrow_struct(),
